fix: rename app/utils/url-regex to app/utils/url.js
